video: tegra: Check force update in fb pan display
Now we'll not update the dc if the x & y offset are not changed
in fb pan display.

But this doesn't work for console switch(e.g: switch from a X server
to a fbcon). The x/y offset will not change but we still need to
update the dc to show the framebuffer console.

So add one more check here: if the update is forced, we'll keep
going to update the dc.

bug 1259633

Change-Id: Ibbb1216151ba34729e438ada62f90eba38758088
Signed-off-by: Mark Zhang <markz@nvidia.com>
Reviewed-on: http://git-master/r/234082
Reviewed-by: Pankaj Dabade <pdabade@nvidia.com>
Tested-by: Pankaj Dabade <pdabade@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Jon Mayo <jmayo@nvidia.com>