power: sbs-battery: Shift work out of get_property function
The board is made to stay awake as long as the power_supply_changed
work thread is in execution. Hence, doing too much work in this thread
often hampers the board from entering LP0 state. Hence, shifting all
work to calculate battery parameters out of this thread and keeping
this thread minimal as to only assign the calculated values. The
calculations needed to obtain the values are not done in the
get_property function anymore.

Bug 1223793

Change-Id: I2595919c9dfd10b0f71136a25f1096a92bba5b3d
Signed-off-by: Darbha Sriharsha <dsriharsha@nvidia.com>
Reviewed-on: http://git-master/r/249641
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>