mmc: host: Fix soft lockup issue
This patch fixes the soft lockup that could occur as a result of commit
8c226304. The mutex_unlock call from interrupt context could cause a
deadlock.

This patch partially reverts the commit 8c226304 by removing the mutex
lock/unlock calls before and after data transfers. mutex locks in
tuning implementation are still present.

This patch ensures that data transfers on all controllers triggered
while running tuning will be stopped until tuning is done inorder to
avoid the risk of any failures due to core voltage changes in tuning.

Locking needs to be extended for the transfer that is already in progress.
This will be done in subsequent patches after thorough testing.

Bug 1354967

Change-Id: Ic6bfd22ec84f24e8782d7da6e3267a7300cb267e
Signed-off-by: Pavan Kunapuli <pkunapuli@nvidia.com>
Reviewed-on: http://git-master/r/266802
Reviewed-by: Mrutyunjay Sawant <msawant@nvidia.com>
Tested-by: Mrutyunjay Sawant <msawant@nvidia.com>